+++ src/FoodService/FoodService.Application/UseCases/CommandHandlers/Product/DeleteProductHandler.cs
@@ -1,60 +1,49 @@
 ﻿using FoodService.Application.Exceptions;
-using FoodService.Domain.Interfaces;
 using FoodService.Application.UseCases.Commands.Product;
+using FoodService.Domain.Entities;
+using FoodService.Domain.Interfaces;
 
-namespace FoodService.Application.UseCases.CommandHandlers.Product
+namespace FoodService.Application.UseCases.CommandHandlers.Product;
+
+public class DeleteProductHandler : ICommandHandler<DeleteProductCommand>
 {
-    public class DeleteProductHandler : ICommandHandler<DeleteProductCommand>
+    private readonly IUnitOfWork _unitOfWork;
+
+    public DeleteProductHandler(IUnitOfWork unitOfWork)
     {
-        private readonly IUnitOfWork _unitOfWork;
+        _unitOfWork = unitOfWork;
+    }
 
-        public DeleteProductHandler(IUnitOfWork unitOfWork)
+    public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
+    {
+        var product = await _unitOfWork.ProductRepository.GetByIdAsync(request.ProductId);
+
+        if (product == null)
         {
-            _unitOfWork = unitOfWork;
+            throw new NotFound("Product not found");
         }
 
-        public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
+        if (product.UserId != request.UserId)
         {
-            var product = await _unitOfWork.ProductRepository.GetByIdAsync(request.ProductId);
-
-            if (product == null)
-            {
-                throw new NotFound("Product not found");
-            }
-
-            if (product.UserId != request.UserId)
-            {
-                throw new Forbidden("You dont have access to this product");
-            }
-
-            var dishes = await _unitOfWork.DishRepository.GetAllAsync(product.UserId);
-
-            //var doesDishContains = dishes != null ? dishes.Any(d => d.Ingredients.Any(i => i.ProductId == product.Id)) : false;
-
-            //if (doesDishContains)
-            //{
-            //    throw new BadRequest("This product there is in some dish");
-            //}
-
-            var dayResults = await _unitOfWork.DayResultRepository.GetAllAsync(product.UserId);
-
-            foreach (var day in dayResults) 
-            {
-                foreach (var meal in day.Meals)
-                { 
-                    foreach (var food in meal.Products)
-                    {
-                        if (food.FoodId == product.Id)
-                        {
-                            throw new BadRequest("This product there is in some meals");
-                        }
-                    }
-                }
-            }
-
-            _unitOfWork.ProductRepository.Delete(product);
-
-            await _unitOfWork.SaveChangesAsync();
+            throw new Forbidden("You dont have access to this product");
         }
+
+        var doesDishContains = await _unitOfWork.RecipeRepository.DoesAnyRecipeContainsProductByIdAsync(product.Id);
+
+        if (doesDishContains)
+        {
+            throw new BadRequest("This product there is in some dish");
+        }
+
+        var doesAnyDayResultContainsProduct = await _unitOfWork.DayResultRepository.DoesAnyDayResultContainsFoodByIdAsync(product.Id, aboutProduct: true);
+
+        if (doesAnyDayResultContainsProduct)
+        {
+            throw new BadRequest("Some day results alredy contain this product");
+        }
+
+        _unitOfWork.ProductRepository.Delete(product);
+
+        await _unitOfWork.SaveChangesAsync();
     }
 }
